+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.2
+
+* Added [ElasticSort] to sort document by a field rather than by score
+* `score` in [ElasticResultMeta] is now optional because of new feature above
+
 ## 0.2.1
 
 * Fields containing "_" were not returned